move block selector to its own file
