+++ src/CMakeLists.txt
@@ -209,6 +209,7 @@ add_library(parthenon
   outputs/restart_hdf5.cpp
   outputs/restart_hdf5.hpp
 
+  pack/block_selector.hpp
   pack/make_pack_descriptor.hpp
   pack/make_swarm_pack_descriptor.hpp
   pack/pack_utils.hpp
